requirements.txt
openai
python-docx
pyannote.audio
torchaudio
ffmpeg-python
matplotlib